+++ android/eir/src/main/java/org/smartregister/fhircore/eir/ui/vaccine/RecordVaccineViewModel.kt
@@ -28,17 +28,15 @@ import org.hl7.fhir.r4.model.Resource
 import org.smartregister.fhircore.eir.data.PatientRepository
 import org.smartregister.fhircore.eir.data.model.PatientVaccineSummary
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity
-import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireConfig
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireViewModel
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 
 class RecordVaccineViewModel(
   application: Application,
-  questionnaireConfig: QuestionnaireConfig,
   val patientRepository: PatientRepository,
   val dispatcherProvider: DispatcherProvider = DefaultDispatcherProvider
-) : QuestionnaireViewModel(application, questionnaireConfig) {
+) : QuestionnaireViewModel(application) {
 
   fun getVaccineSummary(logicalId: String): LiveData<PatientVaccineSummary> {
     val mutableLiveData: MutableLiveData<PatientVaccineSummary> = MutableLiveData()