+++ android/quest/src/test/java/org/smartregister/fhirecore/quest/data/PatientRepositoryTest.kt
@@ -17,26 +17,35 @@
 package org.smartregister.fhirecore.quest.data
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import androidx.lifecycle.MutableLiveData
+import androidx.test.core.app.ApplicationProvider
 import com.google.android.fhir.FhirEngine
-import com.google.android.fhir.FhirEngineProvider.fhirEngine
 import io.mockk.coEvery
+import io.mockk.coVerify
 import io.mockk.mockk
 import java.util.Calendar
 import java.util.Date
 import kotlinx.coroutines.test.runBlockingTest
+import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.DateType
-import org.hl7.fhir.r4.model.Narrative
+import org.hl7.fhir.r4.model.Enumerations
+import org.hl7.fhir.r4.model.Meta
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.hl7.fhir.r4.model.StringType
+import org.hl7.fhir.r4.model.codesystems.AdministrativeGender
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.robolectric.annotation.Config
 import org.smartregister.fhircore.anc.coroutine.CoroutineTestRule
+import org.smartregister.fhircore.engine.configuration.view.SearchFilter
+import org.smartregister.fhircore.engine.configuration.view.registerViewConfigurationOf
+import org.smartregister.fhircore.quest.QuestApplication
 import org.smartregister.fhircore.quest.data.patient.PatientRepository
+import org.smartregister.fhircore.quest.data.patient.model.genderFull
 import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
 import org.smartregister.fhirecore.quest.robolectric.RobolectricTest
 import org.smartregister.fhirecore.quest.shadow.QuestApplicationShadow
@@ -54,8 +63,19 @@ class PatientRepositoryTest : RobolectricTest() {
   fun setUp() {
     fhirEngine = mockk()
 
+    val config =
+      ApplicationProvider.getApplicationContext<QuestApplication>()
+        .registerViewConfigurationOf(
+          primaryFilter = SearchFilter("_tag", "1111", "http://mysystem")
+        )
+
     repository =
-      PatientRepository(fhirEngine, PatientItemMapper, coroutinesTestRule.testDispatcherProvider)
+      PatientRepository(
+        fhirEngine,
+        PatientItemMapper,
+        MutableLiveData(config),
+        coroutinesTestRule.testDispatcherProvider
+      )
   }
 
   @Test
@@ -71,7 +91,7 @@ class PatientRepositoryTest : RobolectricTest() {
   @Test
   fun testLoadDataShouldReturnPatientItemList() = runBlockingTest {
     coEvery { fhirEngine.search<Patient>(any()) } returns
-      listOf(buildPatient("1234", "Doe", "John", 1))
+      listOf(buildPatient("1234", "Doe", "John", 1, Enumerations.AdministrativeGender.FEMALE))
     coEvery { fhirEngine.count(any()) } returns 1
 
     val data = repository.loadData("", 0, true)
@@ -79,16 +99,24 @@ class PatientRepositoryTest : RobolectricTest() {
     Assert.assertEquals("1234", data[0].id)
     Assert.assertEquals("John Doe", data[0].name)
     Assert.assertEquals("1", data[0].age)
+    Assert.assertEquals("F", data[0].gender)
+    Assert.assertEquals("Female", data[0].genderFull())
+
+    coVerify { fhirEngine.search<Patient>(any()) }
   }
 
   @Test
   fun testFetchTestResultsShouldReturnListOfTestReports() {
 
     coEvery { fhirEngine.search<QuestionnaireResponse>(any()) } returns
-      listOf(QuestionnaireResponse().apply { text = Narrative().apply { id = "Blood Count" } })
+      listOf(
+        QuestionnaireResponse().apply {
+          meta = Meta().apply { tag = listOf(Coding().apply { display = "Blood Count" }) }
+        }
+      )
 
     val results = repository.fetchTestResults("1").value
-    Assert.assertEquals("Blood Count", results?.first()?.code?.text)
+    Assert.assertEquals("Blood Count", results?.first()?.meta?.tagFirstRep?.display)
   }
 
   @Test
@@ -109,7 +137,13 @@ class PatientRepositoryTest : RobolectricTest() {
     }
   }
 
-  private fun buildPatient(id: String, family: String, given: String, age: Int): Patient {
+  private fun buildPatient(
+    id: String,
+    family: String,
+    given: String,
+    age: Int,
+    gender: Enumerations.AdministrativeGender = Enumerations.AdministrativeGender.MALE
+  ): Patient {
     return Patient().apply {
       this.id = id
       this.identifierFirstRep.value = id
@@ -117,6 +151,7 @@ class PatientRepositoryTest : RobolectricTest() {
         this.family = family
         this.given.add(StringType(given))
       }
+      this.gender = gender
       this.birthDate = DateType(Date()).apply { add(Calendar.YEAR, -age) }.dateTimeValue().value
 
       this.addAddress().apply {