+++ CHANGELOG.md
@@ -8,9 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2021-xx-xx
 
 ### Added
-
-- Added measure reporting to ANC application
-- Added class for Measure report evaluation which will be used in ANC application
+- ANC | Added progerss bar to cql and measure reporting in ANC patien details page
+- ANC |Added measure reporting to ANC application
+- Engine |Added class for Measure report evaluation which will be used in ANC application
 - ANC | Added Condition resource to sync params list
 - Moved Token to secure storage from AccountManager
 - Expose [custom user attribute](https://www.keycloak.org/docs/latest/server_admin/index.html#_user-attributes) `questionnaire_publisher` available in SharedPreferences with key `USER_QUESTIONNAIRE_PUBLISHER` (#607)
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Patient List, Load Config from server
 - Quest | Added Patient Profile View
 - Quest | Patient Registration Questionnaire
+- Quest | Test Results Questionnaire
+- Quest | Feedback on UI
+- Quest | Remove custom activity
+- Quest | Add tags to patient via questionnaire
 
 ### Fixed